Refactored compartmental models submodule. Moved into new folder.
