+++ covid19_inference/model/compartmental_models/__init__.py
@@ -1,0 +1,10 @@
+from .SIR import SIR
+from .SEIR import SEIR
+from .kernelized_spread import kernelized_spread
+from .uncorrelated_priors import uncorrelated_prior_E, uncorrelated_prior_I
+from .legacy import (
+    SIR_variants,
+    kernelized_spread_variants,
+    kernelized_spread_gender,
+    kernelized_spread_with_interaction,
+)
